task/requirements.txt
amqp==2.3.2
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.4
django-celery-results==1.0.4
django-filter==2.0.0
django-rest-swagger==2.2.0
djangorestframework==3.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10
kombu==4.2.2.post1
MarkupSafe==1.1.0
numpy==1.15.4
openapi-codec==1.3.2
pandas==0.23.4
python-dateutil==2.7.5
python-dotenv==0.10.1
pytz==2018.7
requests==2.21.0
rest-pandas==1.0.0
simplejson==3.16.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
vine==1.1.4
